move isolate.create_sequence_from_records()
